+++ CHANGELOG.md
@@ -13,6 +13,9 @@ When updating the changelog, remember to be very clear about what behavior has c
 and what APIs have changed, if applicable.
 
 ## [Unreleased]
+- Update Gradle version compatibility in the `pegasus` Gradle plugin.
+  - Minimum required Gradle version is now `5.2.1`
+  - Minimum suggested Gradle version is now `5.3`
 
 ## [29.22.14] - 2021-11-24
 - Fix bug where content type was not set for stream requests